+++ Frontend/package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@next/font": "^14.2.15",
     "antd": "^5.23.1",
     "google-fonts": "^1.0.0",
     "next": "15.1.4",
