--- Frontend/src/utils/VotingSystem.js
@@ -1,4 +1,0 @@
-import contractData from "./VotingSystem.json";
-
-export const VotingSystemABI = contractData.abi;
-export const VotingSystemAddress = "0x1F97dCA1FEF579e80d3273229A6Be7C6dC5B9a25";
